stats show collect running and frozen jupyters

+++ django/stats/monitor.py
@@ -25,6 +25,16 @@ It will spot and kill jupyter instances that have had no recent activity
   is also updated with a 'killing' line
 """
 
+class CourseFigures:
+    def __init__(self):
+        self.frozen = 0
+        self.running = 0
+    def count_container(running: bool):
+        if running:
+            self.running +=1
+        else:
+            self.frozen += 1
+
 def monitor(grace: int):
     """
     grace is expressed in minutes
@@ -36,42 +46,51 @@ def monitor(grace: int):
 
     root = nbhosting_settings['root']
     proxy = docker.from_env()
-    containers = proxy.containers.list()
+    containers = proxy.containers.list(all=True)
 
-    count_per_course = {}
+    figures_per_course = {}
     
     for container in containers:
         try:
             name = container.name
             course, student = name.split('-x-')
-            count_per_course.setdefault(course, 0)
+            figures_per_course.setdefault(course, CourseFigures())
+            # stopped containers are useful only for statistics
+            if container.status != 'running':
+                figures_per_course[course].count_container(False)
+                continue
             stamp = os.path.join(root, "students", student, course, ".monitor")
             mtime = os.stat(stamp).st_mtime
             if mtime > grace_past:
-                count_per_course[course] += 1
+                figures_per_course[course] += 1
                 logger.info("sparing {} that had activity {}' ago"
                             .format(name, (now-mtime)//60))
+                figures_per_course[course].count_container(True)
                 continue
             logger.info("Killing container {}".format(name))
             # not sure what signal would be best ?
             container.kill()
+            figures_per_course[course].count_container(False)
             Stats(course).record_kill_jupyter(student)
 
         except FileNotFoundError as e:
             logger.info("container {} has no .monitor stamp - ignoring"
                         .format(name))
+            figures_per_course[course].count_container(container.status == 'running')
+        # typically non-nbhosting continers
         except ValueError as e:
             logger.info("ignoring non-nbhosting container {}"
                         .format(name))
-    for course, count in count_per_course.items():
-        Stats(course).record_jupyters_count(count)
+            # cannot count this one as we don't eveb know in what course it is
+    for course, figures in figures_per_course.items():
+        Stats(course).record_jupyters_count(figures.running, figures.frozen)
         
 def main():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     # default is 2 hours
     parser.add_argument("-g", "--grace", default=2*60, type=int,
                         help="grace timeout in minutes")
-    parser.add_argument("-p", "--period", default=10, type=int,
+    parser.add_argument("-p", "--period", default=15, type=int,
                         help="monitor period in minutes - how often are checked done")
     args = parser.parse_args()
     while True:
